feat: 打包成功1

+++ app/controllers/bookmarks_controller.ts
@@ -2,13 +2,14 @@
  * @Author: lkw199711 lkw199711@163.com
  * @Date: 2024-08-03 05:28:15
  * @LastEditors: lkw199711 lkw199711@163.com
- * @LastEditTime: 2024-08-07 00:40:05
+ * @LastEditTime: 2024-08-11 13:57:32
  * @FilePath: \smanga-adonis\app\controllers\bookmarks_controller.ts
  */
 import type { HttpContext } from '@adonisjs/core/http'
-import type { HttpContextWithUserId } from '#type/http.js'
 import prisma from '#start/prisma'
 import { ListResponse, SResponse } from '../interfaces/response.js'
+import { compressImageToSize } from '../utils/sharp.js'
+import { bookmark_path, get_config, s_delete } from '../utils/index.js'
 export default class BookmarksController {
   public async index({ request, response }: HttpContext) {
     const { chapterId, page, pageSize } = request.only(['page', 'pageSize', 'chapterId', 'order'])
@@ -90,8 +91,9 @@ export default class BookmarksController {
     return response.json(showResponse)
   }
 
-  public async create({ request, response }: HttpContextWithUserId) {
-    const insertData = request.only([
+  public async create({ request, response }: HttpContext) {
+    const { userId } = request as any
+    const { chapterId, mangaId, mediaId, browseType, page, pageImage } = request.only([
       'chapterId',
       'mangaId',
       'mediaId',
@@ -99,9 +101,28 @@ export default class BookmarksController {
       'page',
       'pageImage',
     ])
+
+    // 复制书签图片
+    let outputFile = ''
+    if (pageImage) {
+      const bookmarkPath = bookmark_path()
+      const config = get_config()
+      outputFile = `${bookmarkPath}/smanga_bookmark_${chapterId}.jpg`
+      await compressImageToSize(pageImage, outputFile, config.compress.bookmark)
+    }
+
     const bookmark = await prisma.bookmark.create({
-      data: { ...insertData, userId: request.userId },
+      data: {
+        chapterId,
+        mangaId,
+        mediaId,
+        browseType,
+        page,
+        pageImage: outputFile,
+        userId,
+      },
     })
+
     const saveResponse = new SResponse({ code: 0, message: '新增成功', data: bookmark })
     return response.json(saveResponse)
   }
@@ -120,9 +141,18 @@ export default class BookmarksController {
 
   public async destroy({ params, response }: HttpContext) {
     let { bookmarkId } = params
-    bookmarkId = Number(bookmarkId)
-    const bookmark = await prisma.bookmark.delete({ where: { bookmarkId } })
-    const destroyResponse = new SResponse({ code: 0, message: '删除成功', data: bookmark })
+    const bookmark = await prisma.bookmark.findFirst({ where: { bookmarkId } })
+    if (!bookmark) {
+      return response.json(new SResponse({ code: 1, message: '书签不存在' }))
+    }
+
+    // 删除书签文件
+    if (bookmark.pageImage && /smanga_bookmark/.test(bookmark.pageImage)) {
+      s_delete(bookmark.pageImage)
+    }
+
+    const bookmarkDelete = await prisma.bookmark.delete({ where: { bookmarkId } })
+    const destroyResponse = new SResponse({ code: 0, message: '删除成功', data: bookmarkDelete })
     return response.json(destroyResponse)
   }
 }
